Moved test for time series protocol to own file.
